# Fix query results floating point rounding stability