I opened this PR to have a conversation whether Prometheus should guarantee or not query results stability (floating point rounding) for the same exact query executed multiple times.

## The problem

While comparing query results differences between Cortex and Prometheus for the same exact data and query, I've noticed differences in floating point rounding for some queries. I digged deeper and I've noticed that, for some queries, the same exact query (eg. `curl` command) executed twice on Prometheus returns different results with regards to floating point precision.

This is caused by the (known) fact that PromQL engine internally doesn't guarantee stability in the series order. However, when you sum up a list of floating point values in an order and then you sum up the same exact values but in a different order, the result is not guaranteed to be the exact same.

## Do we wanna solve it?

Should Prometheus guarantee the same exact result for the same exact query run twice on the same exact set of data or no?

_I don't know the answer._

## The fix in this PR

It's very likely that the fix in this PR is incomplete (not every case where series are shuffled has been spotted) but first of all I would like to understand if we should fix it or not.

The main problem I see is having to call `Select()` with `sorted = true` which may cause a significant performance penalty. I believe that within PromQL engine we could fix it without significant performance penalties (eg. issues caused by the usage of `map` could be fixed using `slice` + `map` as done in the couple of places fixed in this PR).

## Benchmarks

### BenchmarkRangeQuery

```
benchmark                                                                                                              old ns/op      new ns/op      delta
BenchmarkRangeQuery/expr=a_one,steps=1-12                                                                              11825          12343          +4.38%
BenchmarkRangeQuery/expr=a_one,steps=10-12                                                                             12154          12708          +4.56%
BenchmarkRangeQuery/expr=a_one,steps=100-12                                                                            19120          19427          +1.61%
BenchmarkRangeQuery/expr=a_one,steps=1000-12                                                                           67936          68148          +0.31%
BenchmarkRangeQuery/expr=a_ten,steps=1-12                                                                              51189          50091          -2.14%
BenchmarkRangeQuery/expr=a_ten,steps=10-12                                                                             54475          54528          +0.10%
BenchmarkRangeQuery/expr=a_ten,steps=100-12                                                                            120777         119286         -1.23%
BenchmarkRangeQuery/expr=a_ten,steps=1000-12                                                                           603249         603562         +0.05%
BenchmarkRangeQuery/expr=a_hundred,steps=1-12                                                                          439457         453718         +3.25%
BenchmarkRangeQuery/expr=a_hundred,steps=10-12                                                                         489319         496974         +1.56%
BenchmarkRangeQuery/expr=a_hundred,steps=100-12                                                                        1202771        1148530        -4.51%
BenchmarkRangeQuery/expr=a_hundred,steps=1000-12                                                                       6055289        5974496        -1.33%
BenchmarkRangeQuery/expr=rate(a_one[1m]),steps=1-12                                                                    15817          16481          +4.20%
BenchmarkRangeQuery/expr=rate(a_one[1m]),steps=10-12                                                                   17315          17638          +1.87%
BenchmarkRangeQuery/expr=rate(a_one[1m]),steps=100-12                                                                  31558          31418          -0.44%
BenchmarkRangeQuery/expr=rate(a_one[1m]),steps=1000-12                                                                 149990         151303         +0.88%
BenchmarkRangeQuery/expr=rate(a_ten[1m]),steps=1-12                                                                    53600          53855          +0.48%
BenchmarkRangeQuery/expr=rate(a_ten[1m]),steps=10-12                                                                   64499          65088          +0.91%
BenchmarkRangeQuery/expr=rate(a_ten[1m]),steps=100-12                                                                  205321         200970         -2.12%
BenchmarkRangeQuery/expr=rate(a_ten[1m]),steps=1000-12                                                                 1416149        1359696        -3.99%
BenchmarkRangeQuery/expr=rate(a_hundred[1m]),steps=1-12                                                                437021         445626         +1.97%
BenchmarkRangeQuery/expr=rate(a_hundred[1m]),steps=10-12                                                               551180         552156         +0.18%
BenchmarkRangeQuery/expr=rate(a_hundred[1m]),steps=100-12                                                              1970910        1927958        -2.18%
BenchmarkRangeQuery/expr=rate(a_hundred[1m]),steps=1000-12                                                             13974374       13576775       -2.85%
BenchmarkRangeQuery/expr=rate(a_one[1m]),steps=10000-12                                                                1934558        1905669        -1.49%
BenchmarkRangeQuery/expr=rate(a_ten[1m]),steps=10000-12                                                                19365799       19027998       -1.74%
BenchmarkRangeQuery/expr=rate(a_hundred[1m]),steps=10000-12                                                            193137059      189064898      -2.11%
BenchmarkRangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=1-12                                                  767917         765364         -0.33%
BenchmarkRangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=10-12                                                 1251542        1284709        +2.65%
BenchmarkRangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=100-12                                                6155303        6095555        -0.97%
BenchmarkRangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=1000-12                                               54993088       54783934       -0.38%
BenchmarkRangeQuery/expr=holt_winters(a_ten[1d],_0.3,_0.3),steps=1-12                                                  6414718        6131718        -4.41%
BenchmarkRangeQuery/expr=holt_winters(a_ten[1d],_0.3,_0.3),steps=10-12                                                 11162729       10993835       -1.51%
BenchmarkRangeQuery/expr=holt_winters(a_ten[1d],_0.3,_0.3),steps=100-12                                                59285826       59767332       +0.81%
BenchmarkRangeQuery/expr=holt_winters(a_ten[1d],_0.3,_0.3),steps=1000-12                                               553900232      542487606      -2.06%
BenchmarkRangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=1-12                                              60816774       60593971       -0.37%
BenchmarkRangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=10-12                                             110285389      109445776      -0.76%
BenchmarkRangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=100-12                                            604520050      594681479      -1.63%
BenchmarkRangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=1000-12                                           5487069073     5411334051     -1.38%
BenchmarkRangeQuery/expr=changes(a_one[1d]),steps=1-12                                                                 601080         601767         +0.11%
BenchmarkRangeQuery/expr=changes(a_one[1d]),steps=10-12                                                                718083         734314         +2.26%
BenchmarkRangeQuery/expr=changes(a_one[1d]),steps=100-12                                                               1842892        2020893        +9.66%
BenchmarkRangeQuery/expr=changes(a_one[1d]),steps=1000-12                                                              13467301       15159101       +12.56%
BenchmarkRangeQuery/expr=changes(a_ten[1d]),steps=1-12                                                                 5282199        5243473        -0.73%
BenchmarkRangeQuery/expr=changes(a_ten[1d]),steps=10-12                                                                6514000        6524001        +0.15%
BenchmarkRangeQuery/expr=changes(a_ten[1d]),steps=100-12                                                               18021299       19676355       +9.18%
BenchmarkRangeQuery/expr=changes(a_ten[1d]),steps=1000-12                                                              138126770      149724644      +8.40%
BenchmarkRangeQuery/expr=changes(a_hundred[1d]),steps=1-12                                                             53935120       52375731       -2.89%
BenchmarkRangeQuery/expr=changes(a_hundred[1d]),steps=10-12                                                            65356349       66618276       +1.93%
BenchmarkRangeQuery/expr=changes(a_hundred[1d]),steps=100-12                                                           183120378      194577137      +6.26%
BenchmarkRangeQuery/expr=changes(a_hundred[1d]),steps=1000-12                                                          1361457725     1482127359     +8.86%
BenchmarkRangeQuery/expr=rate(a_one[1d]),steps=1-12                                                                    600363         590172         -1.70%
BenchmarkRangeQuery/expr=rate(a_one[1d]),steps=10-12                                                                   711596         688125         -3.30%
BenchmarkRangeQuery/expr=rate(a_one[1d]),steps=100-12                                                                  1785140        1561803        -12.51%
BenchmarkRangeQuery/expr=rate(a_one[1d]),steps=1000-12                                                                 11193175       10460653       -6.54%
BenchmarkRangeQuery/expr=rate(a_ten[1d]),steps=1-12                                                                    5097101        5292417        +3.83%
BenchmarkRangeQuery/expr=rate(a_ten[1d]),steps=10-12                                                                   6080097        6010138        -1.15%
BenchmarkRangeQuery/expr=rate(a_ten[1d]),steps=100-12                                                                  16507799       15119997       -8.41%
BenchmarkRangeQuery/expr=rate(a_ten[1d]),steps=1000-12                                                                 114402070      103671889      -9.38%
BenchmarkRangeQuery/expr=rate(a_hundred[1d]),steps=1-12                                                                53139702       51773327       -2.57%
BenchmarkRangeQuery/expr=rate(a_hundred[1d]),steps=10-12                                                               63932660       60676021       -5.09%
BenchmarkRangeQuery/expr=rate(a_hundred[1d]),steps=100-12                                                              216789120      149601935      -30.99%
BenchmarkRangeQuery/expr=rate(a_hundred[1d]),steps=1000-12                                                             1167845312     1035256641     -11.35%
BenchmarkRangeQuery/expr=absent_over_time(a_one[1d]),steps=1-12                                                        579003         577194         -0.31%
BenchmarkRangeQuery/expr=absent_over_time(a_one[1d]),steps=10-12                                                       595453         596583         +0.19%
BenchmarkRangeQuery/expr=absent_over_time(a_one[1d]),steps=100-12                                                      857851         770980         -10.13%
BenchmarkRangeQuery/expr=absent_over_time(a_one[1d]),steps=1000-12                                                     2686190        2666445        -0.74%
BenchmarkRangeQuery/expr=absent_over_time(a_ten[1d]),steps=1-12                                                        4941977        4975842        +0.69%
BenchmarkRangeQuery/expr=absent_over_time(a_ten[1d]),steps=10-12                                                       5144968        5186001        +0.80%
BenchmarkRangeQuery/expr=absent_over_time(a_ten[1d]),steps=100-12                                                      6918800        7038313        +1.73%
BenchmarkRangeQuery/expr=absent_over_time(a_ten[1d]),steps=1000-12                                                     25832776       24513948       -5.11%
BenchmarkRangeQuery/expr=absent_over_time(a_hundred[1d]),steps=1-12                                                    49183748       50497067       +2.67%
BenchmarkRangeQuery/expr=absent_over_time(a_hundred[1d]),steps=10-12                                                   51854039       51659027       -0.38%
BenchmarkRangeQuery/expr=absent_over_time(a_hundred[1d]),steps=100-12                                                  69834460       70375529       +0.77%
BenchmarkRangeQuery/expr=absent_over_time(a_hundred[1d]),steps=1000-12                                                 268388306      240288165      -10.47%
BenchmarkRangeQuery/expr=-a_one,steps=1-12                                                                             12953          13649          +5.37%
BenchmarkRangeQuery/expr=-a_one,steps=10-12                                                                            13897          14069          +1.24%
BenchmarkRangeQuery/expr=-a_one,steps=100-12                                                                           20315          21285          +4.77%
BenchmarkRangeQuery/expr=-a_one,steps=1000-12                                                                          69924          70737          +1.16%
BenchmarkRangeQuery/expr=-a_ten,steps=1-12                                                                             59919          54578          -8.91%
BenchmarkRangeQuery/expr=-a_ten,steps=10-12                                                                            59159          59017          -0.24%
BenchmarkRangeQuery/expr=-a_ten,steps=100-12                                                                           126156         124311         -1.46%
BenchmarkRangeQuery/expr=-a_ten,steps=1000-12                                                                          633184         617758         -2.44%
BenchmarkRangeQuery/expr=-a_hundred,steps=1-12                                                                         545885         481803         -11.74%
BenchmarkRangeQuery/expr=-a_hundred,steps=10-12                                                                        567643         527454         -7.08%
BenchmarkRangeQuery/expr=-a_hundred,steps=100-12                                                                       1333134        1182570        -11.29%
BenchmarkRangeQuery/expr=-a_hundred,steps=1000-12                                                                      6336028        6133039        -3.20%
BenchmarkRangeQuery/expr=a_one_-_b_one,steps=1-12                                                                      23856          24092          +0.99%
BenchmarkRangeQuery/expr=a_one_-_b_one,steps=10-12                                                                     30421          29925          -1.63%
BenchmarkRangeQuery/expr=a_one_-_b_one,steps=100-12                                                                    92023          92706          +0.74%
BenchmarkRangeQuery/expr=a_one_-_b_one,steps=1000-12                                                                   722390         675572         -6.48%
BenchmarkRangeQuery/expr=a_ten_-_b_ten,steps=1-12                                                                      120449         122128         +1.39%
BenchmarkRangeQuery/expr=a_ten_-_b_ten,steps=10-12                                                                     182339         179979         -1.29%
BenchmarkRangeQuery/expr=a_ten_-_b_ten,steps=100-12                                                                    844090         794621         -5.86%
BenchmarkRangeQuery/expr=a_ten_-_b_ten,steps=1000-12                                                                   6676400        6618672        -0.86%
BenchmarkRangeQuery/expr=a_hundred_-_b_hundred,steps=1-12                                                              1100793        1146179        +4.12%
BenchmarkRangeQuery/expr=a_hundred_-_b_hundred,steps=10-12                                                             1726708        1782044        +3.20%
BenchmarkRangeQuery/expr=a_hundred_-_b_hundred,steps=100-12                                                            8525946        8825535        +3.51%
BenchmarkRangeQuery/expr=a_hundred_-_b_hundred,steps=1000-12                                                           79698200       73335379       -7.98%
BenchmarkRangeQuery/expr=a_one_-_b_one,steps=10000-12                                                                  8131451        7710610        -5.18%
BenchmarkRangeQuery/expr=a_ten_-_b_ten,steps=10000-12                                                                  78205511       77308814       -1.15%
BenchmarkRangeQuery/expr=a_hundred_-_b_hundred,steps=10000-12                                                          875727182      844345110      -3.58%
BenchmarkRangeQuery/expr=a_one_and_b_one{l=~'.*[0-4]$'},steps=1-12                                                     53588          54467          +1.64%
BenchmarkRangeQuery/expr=a_one_and_b_one{l=~'.*[0-4]$'},steps=10-12                                                    56075          57400          +2.36%
BenchmarkRangeQuery/expr=a_one_and_b_one{l=~'.*[0-4]$'},steps=100-12                                                   88628          87390          -1.40%
BenchmarkRangeQuery/expr=a_one_and_b_one{l=~'.*[0-4]$'},steps=1000-12                                                  362451         362044         -0.11%
BenchmarkRangeQuery/expr=a_ten_and_b_ten{l=~'.*[0-4]$'},steps=1-12                                                     126205         128295         +1.66%
BenchmarkRangeQuery/expr=a_ten_and_b_ten{l=~'.*[0-4]$'},steps=10-12                                                    163179         155280         -4.84%
BenchmarkRangeQuery/expr=a_ten_and_b_ten{l=~'.*[0-4]$'},steps=100-12                                                   483894         465706         -3.76%
BenchmarkRangeQuery/expr=a_ten_and_b_ten{l=~'.*[0-4]$'},steps=1000-12                                                  3380560        3307605        -2.16%
BenchmarkRangeQuery/expr=a_hundred_and_b_hundred{l=~'.*[0-4]$'},steps=1-12                                             828654         840453         +1.42%
BenchmarkRangeQuery/expr=a_hundred_and_b_hundred{l=~'.*[0-4]$'},steps=10-12                                            1173744        1175346        +0.14%
BenchmarkRangeQuery/expr=a_hundred_and_b_hundred{l=~'.*[0-4]$'},steps=100-12                                           5377992        4764257        -11.41%
BenchmarkRangeQuery/expr=a_hundred_and_b_hundred{l=~'.*[0-4]$'},steps=1000-12                                          44327746       38920191       -12.20%
BenchmarkRangeQuery/expr=a_one_or_b_one{l=~'.*[0-4]$'},steps=1-12                                                      54449          54395          -0.10%
…