@brian-brazil 2020-08-06
> The main problem I see is having to call Select() with sorted = true which may cause a significant performance penalty

Indeed, I went to quite a bit of effort to add that optimisation. It really helps with queries on recent high-cardinality data.

Any proposals here should be entirely within PromQL, not touching TSDB stuff (which you'd need anyway for correctness).