@pracucci 2020-08-06
> > The main problem I see is having to call Select() with sorted = true which may cause a significant performance penalty
> 
> Indeed, I went to quite a bit of effort to add that optimisation. It really helps with queries on recent high-cardinality data.

I understand this @brian-brazil. However, I believe we should answer the question whether Prometheus should guarantee query results stability or not. Have you ever discussed it within the community?