@brian-brazil 2020-08-06
It's come up one or twice with `deriv` where it was causing actual problems, but in general it's not a big deal. If there's a cheap way to make it better we'd take it, but fundamentally it doesn't matter if the 7th decimal place isn't precisely right.

Also as I understand it, even with this in you'd still have inconsistencies between Cortex and Prometheus due to splitting computations across machines.