@pracucci 2020-08-06
> Also as I understand it, even with this in you'd still have inconsistencies between Cortex and Prometheus due to splitting computations across machines.

Could you elaborate a bit more on this? Are you talking about the query-frontend time-based (24h) query splitting?