@beorn7 2020-08-06
> If there's a cheap way to make it better we'd take it, but fundamentally it doesn't matter if the 7th decimal place isn't precisely right.

That's essentially my opinion, too. It's convenient for testing and similar purposes if results are as reproducible as possible, but as long as we are just talking about rounding errors and not actually wrong results, it's not worth a noticeable performance penalty or even a penalty in code complexity.