@brian-brazil 2020-08-06
> Are you talking about the query-frontend time-based (24h) query splitting?

No.

> Could you elaborate a bit more on this?

My understanding is that one of the things Cortex is looking at is distributing computation of things like `sum` across machines. Floating point math is not associative, so that can produce very slightly different results.