@bwplotka 2020-08-06
Same opinion on my front as @beorn7 and @brian-brazil - not feeling strongly about that precision, especially if it means much more complex code or slow downs. But if we can have it for cheap, sure!