@pracucci 2020-08-07
Agree with you. Looks a minor thing and the penalty introduced by sorted `Select()` looks not worth. Thanks for the discussion! I'm going to close this PR.

Off topic: introduced tolerance to the Cortex query-tee results comparison ([PR](https://github.com/cortexproject/cortex/pull/2994)).